Fixes to get mac bulld working with Xquartz

+++ linuxdoom-1.10/d_main.c
@@ -1119,7 +1119,8 @@ void D_DoomMain (void)
 	// for statistics driver
 	extern  void*	statcopy;                            
 
-	statcopy = (void*)atoi(myargv[p+1]);
+	// Sam: use atol instead of atoi
+	statcopy = (void*)atol(myargv[p+1]);
 	printf ("External statistics registered.\n");
     }
